refactor: move by-country -> news, renames
